Fix the test for Game#version
I suspect this was only passing before because of how `freeze_time`
works, internally. But Game#version does return a Float now (where it
used to be an Integer)... so this test was out of date.

+++ test/models/game_test.rb
@@ -159,7 +159,7 @@ class GameTest < ActiveSupport::TestCase
 
       it "returns the expected Integer" do
         freeze_time do
-          _(subject.version).must_equal(Time.now.to_i)
+          _(subject.version).must_equal(Time.now.to_f)
         end
       end
     end
